Fix wrong order of callback of sendUserMessage
This caused a promise rejection every time a message was sent, which is usually left unhandled.

This issue was caused by 6f8e856 which changed the order of Sendbird's callbacks for almost everything. The sendUserMessage callback for the send function that is passed to the command handler was not reversed, and later in e0ab6bb this issue was exposed.